feat : 일정 삭제 및 벌크 처리 기능 구현

+++ BE/src/main/java/com/bibbidi/wedding/appointment/service/AppointmentDeleteServiceInterface.java
@@ -3,6 +3,7 @@ package com.bibbidi.wedding.appointment.service;
 import java.util.List;
 
 public interface AppointmentDeleteServiceInterface {
-    void changeAllToNewChecklistItemIds(Long newChecklistItemId, List<Integer> targetAppointmentIds);
-    void deleteAll(List<Integer> targetAppointmentIds);
+    void changeAllToNewChecklistItemIds(Long newChecklistItemId, List<Long> targetAppointmentIds);
+
+    void deleteAll(List<Long> targetAppointmentIds);
 }
